Revert "Ft2 Isolate constrained objective function"
