+++ src/objective/constrained_xentropy_objective.hpp
@@ -18,7 +18,7 @@
 #define LIGHTGBM_OBJECTIVE_CONSTRAINED_XENTROPY_OBJECTIVE_HPP_
 
 #include <LightGBM/meta.h>
-#include <LightGBM/constrained_objective_function.h>
+#include <LightGBM/objective_function.h>
 #include <LightGBM/utils/common.h>
 #include <LightGBM/utils/constrained.hpp>
 #include "../metric/xentropy_metric.hpp"
